backend/requirements.txt
certifi==2025.1.31
fastapi==0.115.6
playwright==1.49.1
python-dotenv==1.0.1
requests==2.32.3
supabase==2.29.0
python-telegram-bot==21.10
uvicorn[standard]==0.34.0
